Add routes for about, registration, login page, also updated App.js to include new routes
