--- client/soundland/src/Components/AboutSoundLand.js
@@ -1,21 +1,0 @@
-import React from 'react';
-import Footer from './Footer';
-//import Nav from './Nav';
-
-    class AboutSongLand {
-
-        render (){
-            return(
-              <div>
-                <div class="aboutapp"><h1>About SoundLand</h1>
-                <p>Looking to make new friends who share the same kind of music you love?  Are you a promoter looking to promote a new artist to your target audience? Or you are just looking for a place on Saturday night that plays your favorite oldies? <strong>SoundLand</strong> is the first ever app to bring local music lovers together.  Search your favorite kind of music and find other people near you that also enjoy the same music.  After a simple registration process, you can find out the latest music events in your area! You can also create your own events and attract new friends to watch your favorite band perform at your local bar!  
-                </p>
-                </div>
-                </ Footer>  
-              </div>
-             
-         );
-        }
-       }
-
-  export default AboutSongLand;
