+++ client/soundland/src/Components/AboutUs.js
@@ -1,37 +1,51 @@
 import React from 'react';
-//import Nav from './Nav';
+import { Link, Redirect } from 'react-router-dom';
+
+import Header from './Header';
 import Footer from './Footer';
-class AboutUs {
+
+class About {
 
   render (){
     return(
-        </ Nav>
-        <div class="aboutus"><h2>About Team Yar-ss-ary-phanie</h2>
-            <div class="yarden"><h3>Yarden Hochman</h3>
-            <img src=""/>
-            <a href="https://github.com/yardenhochman">Yarden's GitHub Page</a>
-            <p></p>
-            </div>
-            <div class="issac"><h3>Issac Grunwald</h3>
-            <img src=""/>
-            <a href="https://github.com/issacg1">Issac's GitHub Page</a>
-            <p></p>
-            </div>
-            <div class="gary"><h3>Gary Chin</h3>
-            <img src=""/>
-            <a href="https://github.com/Gnyc1001">Gary's GitHub Page</a>
-            <p>Born and raised in central NJ.  Graduated from Pace University with a BBA in Finance.</p>
-            </div>
-            <div class="stephanie"><h3>Stephanie Fung</h3>
-            <img src=""/>
-            <a href="https://github.com/stephdfung">Stephanie's GitHub Page</a>
-            <p></p>
-            </div>
-        </div>
+      <div>
+        <Header/>
 
-        </ Footer>
+        <div class="about">
+          <h2>The SoundLand Team</h2>
+
+            <div class="team">
+              <h3>Gary Chin</h3>
+              <img src=""/>
+              <Link to={"https://github.com/Gnyc1001"}>Gary's GitHub Page</Link>
+              <p>Born and raised in central NJ.  Graduated from Pace University with a BBA in Finance.</p>
+            </div>
+
+            <div class="team">
+              <h3>Stephanie Fung</h3>
+              <img src=""/>
+              <Link to={"https://github.com/stephdfung"}>Stephanie's GitHub Page</Link>
+              <p></p>
+            </div>
+
+            </div>
+            <div class="team">
+              <h3>Issac Grunwald</h3>
+              <img src=""/>
+              <Link to={"https://github.com/issacg1"}>Issac's GitHub Page</Link>
+              <p></p>
+            </div>
+
+            <div class="team">
+              <h3>Yarden Hochman</h3>
+              <img src=""/>
+              <Link to={"https://github.com/yardenhochman"}>Yarden's GitHub Page </Link>
+              <p></p>
+            </div>
+        <Footer />
+      </div>
     );
    }
 }
 
-export default AboutUs;
+export default About;
