+++ client/soundland/src/Components/Login.js
@@ -1,15 +1,12 @@
 import React, { Component } from 'react';
 import axios from 'axios';
 
-class Registration extends Component {
+class Login extends Component {
   constructor() {
     super();
     this.state = {
-      name: '',
       username: '',
-      email: '',
       password_digest: '',
-      password_confirm: '',
       fireRedirect: false,
     }
   }
@@ -25,17 +22,12 @@ class Registration extends Component {
 
   handleFormSubmit(event) {
       event.preventDefault();
-    if (this.state.password_digest === this.state.password_confirm) {
-      // fetch POST request to server to create new user
-      // redirect to their profile? back two pages?
 
       let data = {
-        name: this.state.name,
         username: this.state.username,
-        email: this.state.email,
         password_digest: this.password_digest
       }
-//routes here are not valid revist
+        //routes here are not valid revist
       axios({
         method: 'POST',
         url: 'http://localhost:3001/auth/',
@@ -44,54 +36,32 @@ class Registration extends Component {
       .then(res => {
         console.log(res.data);
         this.setState({
-          newID: res.data.id,
+          id: res.data.id,
           //The res.data.id might be wrong here
           fireRedirect: true,
         });
       }).catch(err=> console.log(err));
       event.target.reset();
-
-    } else {
-      alert('Passwords do not match')
-      // we can update the alert later to be more complex
-      this.setState({
-        password_digest: '',
-        password_confirm: ''
-      })
-    }
+    } 
   }
 
   render() {
     return(
-      <div className="register">
+      <div className="login">
 
-        <div className="register-top">
+        <div className="login-top">
           <img className="profile-icon"
           src="https://d30y9cdsu7xlg0.cloudfront.net/png/898318-200.png"/>
-          <h3>Register</h3>
+          <h3>Login</h3>
         </div>
 
-        <div className="register-form">
+        <div className="login-form">
           <form onSubmit={this.handleFormSubmit}>
-            <input
-              type="text"
-              placeholder="Name"
-              name="name"
-              value={this.state.name}
-              onChange={this.handleInputChange}
-            />
             <input
               type="text"
               placeholder="Username"
               name="username"
               value={this.state.username}
-              onChange={this.handleInputChange}
-            />
-            <input
-              type="email"
-              placeholder="Email"
-              name="email"
-              value={this.state.email}
               onChange={this.handleInputChange}
             />
             <input
@@ -103,20 +73,12 @@ class Registration extends Component {
               onChange={this.handleInputChange}
             />
             <input
-              type="password"
-              placeholder="Confirm Password"
-              name="password_confirm"
-              minLength="6" required 
-              value={this.state.password_confirm}
-              onChange={this.handleInputChange}
-            />
-            <input
               type="submit"
               value="Submit"
             />
           </form>
           {this.state.fireRedirect
-          ? <Redirect push to={`/user/${this.state.newId}`} />
+          ? <Redirect push to={`/user/${this.state.id}`} />
           : ''}
         </div>
 
@@ -124,6 +86,4 @@ class Registration extends Component {
     )
   }
 
-}
-
-export default Registration;
+export default Login;
